Completed HW4.

+++ src/homework6/DepositCalculatorToRefactor.java
@@ -1,12 +1,13 @@
 package homework6;
 
+import java.math.BigDecimal;
 import java.util.Scanner;
 
 public class DepositCalculatorToRefactor {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.print("Введите сумму вклада в банк в гривнах: ");
-        double sum = sc.nextDouble();
+        int sum =sc.nextInt();
         System.out.print("Какой процент годовых? ");
         double percent = sc.nextDouble();
         System.out.print("Длительность вклада (лет): ");
@@ -15,7 +16,7 @@ public class DepositCalculatorToRefactor {
 
         for (int i = 1; i <= term; i++) {
             double y = sum / 100 * percent;
-            sum = sum + y;
+            sum = (int)(sum + y);
             System.out.println("Накопленная сумма за " + i + (i % 10 == 1 ? " год" :
                     i % 10 == 2 || i % 10 == 3 || i % 10 == 4 ? " года"
                             : " лет") + " составила: " + String.format("%.2f",sum));
